Default values
Default values are set as follows: 4096 samples, amplitude equal to 1 mm, frequency equal to 1 Hz, and phase angle equal to 0 degrees.
The sampling frequency is removed because the signal is generated by the processor and not obtained externally through a physical measurement.
The presentation of the data in the console and the graph is optimized.